refactored and exposes more utility methods and updated README
