+++ README.md
@@ -3,7 +3,7 @@
 # Decentralized Web Node (DWN) SDK
 
 Code Coverage
-![Statements](https://img.shields.io/badge/statements-94.85%25-brightgreen.svg?style=flat) ![Branches](https://img.shields.io/badge/branches-94.53%25-brightgreen.svg?style=flat) ![Functions](https://img.shields.io/badge/functions-92.04%25-brightgreen.svg?style=flat) ![Lines](https://img.shields.io/badge/lines-94.85%25-brightgreen.svg?style=flat)
+![Statements](https://img.shields.io/badge/statements-94.88%25-brightgreen.svg?style=flat) ![Branches](https://img.shields.io/badge/branches-94.56%25-brightgreen.svg?style=flat) ![Functions](https://img.shields.io/badge/functions-92.13%25-brightgreen.svg?style=flat) ![Lines](https://img.shields.io/badge/lines-94.88%25-brightgreen.svg?style=flat)
 
 ## Introduction
 
@@ -25,11 +25,13 @@ npm install @tbd54566975/dwn-sdk-js
 
 ```javascript
 
-import { Dwn, DataStream, RecordsWrite, RecordsQuery } from '@tbd54566975/dwn-sdk-js';
+import { Dwn, DataStream, DidKeyResolver, Jws, RecordsWrite, RecordsQuery } from '@tbd54566975/dwn-sdk-js';
 
 export const dwn = await Dwn.create({});
 
 ...
+const didKey = await DidKeyResolver.generate(); // generate a did:key DID
+const signatureMaterial = Jws.createSignatureInput(didKey);
 const data = randomBytes(32); // in node.js
 // or in web
 // const data = new Uint8Array(32);
