+++ src/index.ts
@@ -26,9 +26,11 @@ export { Dwn } from './dwn.js';
 export { DwnInterfaceName, DwnMethodName } from './core/message.js';
 export { Encoder } from './utils/encoder.js';
 export { HooksWrite, HooksWriteOptions } from './interfaces/hooks/messages/hooks-write.js';
+export { Jws } from './utils/jws.js';
+export { KeyMaterial, PrivateJwk, PublicJwk } from './jose/types.js';
 export { MessageReply } from './core/message-reply.js';
 export { MessageStore } from './store/message-store.js';
 export { MessageStoreLevel } from './store/message-store-level.js';
-export { PrivateJwk, PublicJwk } from './jose/types.js';
 export { ProtocolsConfigure, ProtocolsConfigureOptions } from './interfaces/protocols/messages/protocols-configure.js';
 export { ProtocolsQuery, ProtocolsQueryOptions } from './interfaces/protocols/messages/protocols-query.js';
+export { SignatureInput } from './jose/jws/general/types.js';
